Set up intro section and add Google fonts for Quizzical App
This commit sets up the intro section of the Quizzical App and adds all the Google font files. It also creates the HTML and CSS components for the intro section. The intro section will provide an introduction to the App and give users a brief overview of the features and capabilities.